Rename openfoam->additivefoam; ignore cases which are existing filepaths
