+++ uq_workflow/Stage1/inputs.json
@@ -1,5 +1,5 @@
 {
-    "openfoam":
+    "additivefoam":
     {
         "inputs":
         [
