+++ uq_workflow/Stage1/parsers/__init__.py
@@ -1,2 +1,2 @@
 from .exaca import create_cases
-from .openfoam import create_cases
+from .additivefoam import create_cases
